.github/actions/validate_pr_title/requirements.txt
atlassian-python-api>=3.41.0
requests>=2.31.0
python-dotenv>=1.0.0
PyGithub>=1.59